send user messages to the user directly, not to daemon
moves most "user" functionality to the test users for tests
